+++ src/LaravelPlans/Contracts/PlanFeatureInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Contracts;
+
+interface PlanFeatureInterface
+{
+    public function plan();
+    public function usage();
+}