+++ src/LaravelPlans/Contracts/PlanInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Contracts;
+
+interface PlanInterface
+{
+    public function features();
+    public function subscriptions();
+    public function isFree();
+    public function hasTrial();
+}