+++ src/LaravelPlans/Contracts/PlanSubscriberInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Contracts;
+
+interface PlanSubscriberInterface
+{
+    public function subscription($name = 'default');
+    public function subscriptions();
+    public function subscribed($subscription = 'default');
+    public function newSubscription($name, $plan);
+    public function subscriptionUsage($subscription = 'default');
+}