+++ src/LaravelPlans/Events/SubscriptionRenewed.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Events;
+
+use Czechbox\LaravelPlans\Models\PlanSubscription;
+use Illuminate\Queue\SerializesModels;
+
+class SubscriptionRenewed
+{
+    use SerializesModels;
+
+    /**
+     * @var \Laraplans\Models\PlanSubscription
+     */
+    public $subscription;
+
+    /**
+     * Create a new event instance.
+     *
+     * @param  \Laraplans\Models\PlanSubscription  $subscription
+     * @return void
+     */
+    public function __construct(PlanSubscription $subscription)
+    {
+        $this->subscription = $subscription;
+    }
+}