+++ src/LaravelPlans/Exceptions/FeatureValueFormatIncompatibleException.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Exceptions;
+
+class FeatureValueFormatIncompatibleException extends \Exception
+{
+    /**
+     * Create a new FeatureValueFormatIncompatibleException instance.
+     *
+     * @param $feature
+     * @return void
+     */
+    public function __construct($value)
+    {
+        $this->message = "Feature value format is incompatible: {$value}.";
+    }
+}