+++ src/LaravelPlans/Exceptions/InvalidPlanFeatureException.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Czechbox\LaravelPlans\Exceptions;
+
+class InvalidPlanFeatureException extends \Exception
+{
+    /**
+     * Create a new InvalidPlanFeatureException instance.
+     *
+     * @param $feature
+     * @return void
+     */
+    public function __construct($feature)
+    {
+        $this->message = "Invalid plan feature: {$feature}";
+    }
+}